update, remove item byte methods
