--- pineapple-bundle/pineapple-nms/spigot-v1_20_R4/src/main/java/sh/miles/pineapple/nms/impl/v1_20_R4/packet/PineapplePacketsImpl.java
@@ -1,15 +1,0 @@
-package sh.miles.pineapple.nms.impl.v1_20_R4.packet;
-
-import org.bukkit.craftbukkit.v1_20_R4.entity.CraftLivingEntity;
-import org.bukkit.entity.LivingEntity;
-import org.jetbrains.annotations.NotNull;
-import sh.miles.pineapple.nms.api.packet.PineapplePackets;
-
-public class PineapplePacketsImpl implements PineapplePackets {
-
-    @Override
-    public void broadcastTotemEffect(@NotNull final LivingEntity entity) {
-        final var livingEntity = ((CraftLivingEntity) entity).getHandle();
-        livingEntity.level().broadcastEntityEvent(livingEntity, (byte) 35);
-    }
-}
