+++ pineapple-bundle/pineapple-nms/spigot-v1_21_R3/src/main/java/sh/miles/pineapple/nms/impl/v1_21_R3/PineappleNMSImpl.java
@@ -1,4 +1,4 @@
-package sh.miles.pineapple.nms.impl.v1_21_R2;
+package sh.miles.pineapple.nms.impl.v1_21_R3;
 
 import com.google.common.base.Preconditions;
 import com.google.common.primitives.Ints;
@@ -6,7 +6,6 @@ import com.mojang.datafixers.DataFixer;
 import com.mojang.serialization.Dynamic;
 import net.md_5.bungee.api.chat.BaseComponent;
 import net.minecraft.SharedConstants;
-import net.minecraft.core.NonNullList;
 import net.minecraft.core.component.DataComponents;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.NbtAccounter;
@@ -18,17 +17,13 @@ import net.minecraft.network.protocol.game.ClientboundOpenScreenPacket;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.util.datafix.fixes.References;
-import net.minecraft.world.ContainerHelper;
 import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.inventory.MenuType;
 import net.minecraft.world.item.component.ItemLore;
-import org.bukkit.Bukkit;
-import org.bukkit.craftbukkit.v1_21_R2.CraftWorld;
-import org.bukkit.craftbukkit.v1_21_R2.entity.CraftPlayer;
-import org.bukkit.craftbukkit.v1_21_R2.event.CraftEventFactory;
-import org.bukkit.craftbukkit.v1_21_R2.inventory.CraftContainer;
-import org.bukkit.craftbukkit.v1_21_R2.inventory.CraftItemStack;
-import org.bukkit.craftbukkit.v1_21_R2.util.CraftMagicNumbers;
+import org.bukkit.craftbukkit.v1_21_R3.entity.CraftPlayer;
+import org.bukkit.craftbukkit.v1_21_R3.event.CraftEventFactory;
+import org.bukkit.craftbukkit.v1_21_R3.inventory.CraftContainer;
+import org.bukkit.craftbukkit.v1_21_R3.inventory.CraftItemStack;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.InventoryView;
@@ -38,8 +33,8 @@ import org.jetbrains.annotations.Nullable;
 import sh.miles.pineapple.ReflectionUtils;
 import sh.miles.pineapple.nms.api.PineappleNMS;
 import sh.miles.pineapple.nms.api.PineappleUnsafe;
-import sh.miles.pineapple.nms.impl.v1_21_R2.internal.ComponentUtils;
-import sh.miles.pineapple.nms.impl.v1_21_R2.packet.PineapplePacketsImpl;
+import sh.miles.pineapple.nms.impl.v1_21_R3.internal.ComponentUtils;
+import sh.miles.pineapple.nms.impl.v1_21_R3.packet.PineapplePacketsImpl;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
